Remove cdi and rename cd2 -> cdi
