--- cdi-reactive-messaging/src/test/java/com/amadeus/middleware/odyssey/reactive/messaging/core/cdiextension/Bar.java
@@ -1,4 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-public class Bar implements AnnotatedInterface {
-}
