--- cdi-reactive-messaging/src/test/java/com/amadeus/middleware/odyssey/reactive/messaging/core/cdiextension/Foo.java
@@ -1,6 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-@MyAnnotation
-@MyInheritedAnnotation
-public class Foo {
-}
