--- cdi-reactive-messaging/src/test/java/com/amadeus/middleware/odyssey/reactive/messaging/core/cdiextension/MyAnnotation.java
@@ -1,11 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface MyAnnotation {
-}
