--- cdi-reactive-messaging/src/test/utils/AnnotatedInterface.java
@@ -1,6 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-@MyAnnotation
-@MyInheritedAnnotation
-public interface AnnotatedInterface {
-}
