--- cdi-reactive-messaging/src/test/utils/Foo.java
@@ -1,6 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-@MyAnnotation
-@MyInheritedAnnotation
-public class Foo {
-}
