--- cdi-reactive-messaging/src/test/utils/MyInheritedAnnotation.java
@@ -1,13 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Inherited
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface MyInheritedAnnotation {
-}
