--- cdi-reactive-messaging/src/test/utils/SubBar.java
@@ -1,4 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-public class SubBar extends Bar {
-}
