--- cdi-reactive-messaging/src/test/utils/SubFoo.java
@@ -1,4 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-public class SubFoo extends Foo {
-}
