--- cdi2-reactive-messaging/src/main/java/com/amadeus/middleware/odyssey/reactive/messaging/core/cdiextension/CDIAsync.java
@@ -1,18 +1,0 @@
-package com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension;
-
-import com.amadeus.middleware.odyssey.reactive.messaging.core.Async;
-
-public class CDIAsync<T> implements Async<T> {
-
-  private Class<T> clazz;
-
-  public CDIAsync(Class<T> clazz) {
-    this.clazz = clazz;
-  }
-
-  @Override
-  public T get() {
-    return MessageScopedContext.getInstance()
-        .get(clazz);
-  }
-}
