--- cdi2-reactive-messaging/src/main/resources/META-INF/beans.xml
@@ -1,5 +1,0 @@
-<beans xmlns="http://java.sun.com/xml/ns/javaee"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-  http://java.sun.com/xml/ns/javaee/beans_1_0.xsd">
-</beans>
