--- cdi2-reactive-messaging/src/main/resources/META-INF/services/com.amadeus.middleware.odyssey.reactive.messaging.core.impl.MessageBuilderProvider
@@ -1,1 +1,0 @@
-com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension.CDIMessageBuilderProvider
