--- cdi2-reactive-messaging/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-com.amadeus.middleware.odyssey.reactive.messaging.core.cdiextension.StreamExtension
