+++ test-cdi-reactive-messaging/pom.xml
@@ -33,7 +33,7 @@
     <dependencies>
         <dependency>
             <groupId>com.amadeus.middleware.odyssey-reactive-messaging</groupId>
-            <artifactId>cdi2-reactive-messaging</artifactId>
+            <artifactId>cdi-reactive-messaging</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
